# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZEROHOUR REALTOR PULSE API",
    "version": "0.1.0",
    "description": "Phase 1 local MVP API contract for ZEROHOUR REALTOR PULSE. This contract supports frontend integration, simulated creator onboarding, async provisioning, mock ZIP signal ingestion, cached dashboards, signal actions, billing handoff, and basic admin operations."
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development API"
    }
  ],
...
  "paths": {
...
    "/signals": {
      "get": {
        "tags": ["signals"],
        "summary": "List active signals",
        "operationId": "listSignals",
        "parameters": [
          {
            "name": "zip_code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[0-9]{5}$"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active signal list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalListResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/signals/resolve-owner": {
      "post": {
        "tags": ["signals"],
        "summary": "Resolve homeowner name and mailing address for one explicit signal-card action",
        "description": "Use only when an agent clicks Send Letter or expands a signal card. This endpoint is rate-limited to 50 calls per agent per day and stores resolved owner data only in temporary session cache.",
        "operationId": "resolveOwner",
        "x-exact-alias": "/api/signals/resolve-owner",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnerResolutionRequest"
              },
              "example": {
                "address": "2218 Coldwater Cyn",
                "zip": "90210"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Owner resolution result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerResolutionResponse"
                },
                "examples": {
                  "resolved": {
                    "value": {
                      "resolved": true,
                      "name": "Taylor Carter",
                      "mailing_address": {
                        "line1": "2218 Coldwater Cyn",
                        "line2": null,
                        "city": "Los Angeles",
                        "state": "CA",
                        "zip": "90210"
                      }
                    }
                  },
                  "unresolved": {
                    "value": {
                      "resolved": false,
                      "name": null,
                      "mailing_address": null
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Owner resolution daily limit exceeded"
          },
          "503": {
            "description": "Rate limiter unavailable"
          }
        }
      }
    },
    "/signals/{signal_id}/view": {
      "post": {
        "tags": ["signals"],
        "summary": "Record signal viewed event",
        "operationId": "viewSignal",
        "parameters": [
          {
            "$ref": "#/components/parameters/SignalId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Status"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/signals/{signal_id}/claim": {
      "post": {
        "tags": ["signals"],
        "summary": "Claim a signal opportunity",
        "operationId": "claimSignal",
        "parameters": [
          {
            "$ref": "#/components/parameters/SignalId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Status"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/signals/{signal_id}/outcomes": {
      "post": {
        "tags": ["signals"],
        "summary": "Record signal performance feedback",
        "operationId": "recordSignalOutcome",
        "parameters": [
          {
            "$ref": "#/components/parameters/SignalId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutcomeRequest"
              },
              "example": {
                "outcome_type": "appointment_booked",
                "properties": {
                  "channel": "phone"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Status"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "SignalId": {
        "name": "signal_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "responses": {
      "Status": {
        "description": "Status response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StatusResponse"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Not found"
      },
...
    },
    "schemas": {
...
      "SignalSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "zip_code": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "score": {
            "type": "number"
          },
          "confidence": {
            "type": "number"
          },
          "freshness_days": {
            "type": "integer"
          },
          "source": {
            "type": "string"
          },
          "address": {
            "type": ["string", "null"]
          },
          "signal_fired_at": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "signal_claimed_at": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "claimed_by_user_id": {
            "type": ["string", "null"],
            "format": "uuid"
          },
          "claim_delta_minutes": {
            "type": ["integer", "null"]
          }
        },
        "additionalProperties": true
      },
...
      "OwnerResolutionRequest": {
        "type": "object",
        "required": ["address", "zip"],
        "properties": {
          "address": {
            "type": "string",
            "minLength": 3
          },
          "zip": {
            "type": "string",
            "minLength": 5,
            "maxLength": 10
          }
        }
      },
      "MailingAddress": {
        "type": "object",
        "required": ["line1"],
        "properties": {
          "line1": {
            "type": "string"
          },
          "line2": {
            "type": ["string", "null"]
          },
          "city": {
            "type": ["string", "null"]
          },
          "state": {
            "type": ["string", "null"]
          },
          "zip": {
            "type": ["string", "null"]
          }
        }
      },
      "OwnerResolutionResponse": {
        "type": "object",
        "required": ["resolved"],
        "properties": {
          "resolved": {
            "type": "boolean"
          },
          "name": {
            "type": ["string", "null"]
          },
          "mailing_address": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/MailingAddress"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "SignalListResponse": {
        "type": "object",
        "required": ["signals"],
        "properties": {
          "signals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SignalSummary"
            }
          }
        }
      },
      "OutcomeRequest": {
        "type": "object",
        "required": ["outcome_type"],
        "properties": {
          "outcome_type": {
            "type": "string",
            "examples": ["contacted", "follow_up_scheduled", "appointment_booked", "listing_won", "deal_closed"]
          },
          "properties": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
      "StatusResponse": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": ["string", "null"]
          }
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}